Bug: M20 - List SD card
When the SD card is inserted in a Mac computer some special directory are written. The purse firmware is unable to navigate those directory and generates an error sent showed on the serial line.

the BUG is the instruction SERIAL_ECHOLN does not work with string pointers and there fore garbage is sent on the serial line.

The proposed correction is to use the right instruction: SERIAL_ECHORPGM that works with string pointers.
Therefore the correct message is MSG_SD_CANT_ENTER_SUBDIR and not the original listed.