G76: fix Z homing after steel sheet removal
When the steel sheet is in place, the user is asked to remove the
steel sheet before continuing the temperature calibration process.
After removal acknowledge from user, a Z homing is performed but world
coordinates are not properly updated and the extruder nozzle can crash
into the bed and then damage the bed and the nozzle.