Added support for the upstream Marlin interpretation of the M204 code:
M204 S.. T..:
T is interpreted the old way (as acceleration when retracting)
only if an S code is found at the same line.
This allows PrusaResearch to interpret the legacy G-codes generated
by our older Slic3r with older Slic3r profiles.

M204 P.. R.. T..:
T is ignored, P is interpreted as acceleration when extruding,
R is interpreted as acceleration when retracting.
This will be the format the Slic3r 1.41.0 will produce from
the Machine Limits page.
In the future both MK3 firmware and Slic3r will likely be extended
to support the separate travel acceleration.

This change will help us to solve the following Slic3r issue:
https://github.com/prusa3d/Slic3r/issues/1089