Make MMUv2 compatible for MK3 Einsy and MK2.5 miniRambo boards.
The defined MMU_RST_PIN pin 76 is only available on the P3 connector on the Einsy board BUT not on the miniRAMbo.

As the MMUv2 should be supported on the MK3 and MK2.5, which has a miniRAMbo controller, you gonna need to use another pin on the miniRAMbo 1.0 and 1.3 for MMU_RST_PIN.

Like 'X-max' is used for Noctua tacho signal I propose to use 'z-max' for MMU_RST_PIN. Prusa also gonna need a split cable for the miniRAMBO MMUv2 as for the MK2.5 Noctuca and PINDAv2.