Drop min nozzle temperature to 5 degrees
This drops the minimum nozzle temperature from 15 degrees celsius to 5 degrees celsius for both the MK3 and MK3S. The 15
degree lower limit was a remnant of the MK2 board which could not sense lower than 15 degrees celsius. The MK3 (and MK3S) can
both go down to 0 degrees, so 5 degrees is a relatively arbitrary amount of safety margin over 0. This value does not come
from the thermistor itself, but from the firmware which has been implemented to not read below 0.