Prepare PF-build.sh for Prusa Arduino_Boards V1.0.2 As I am planning to add new Prusa Arduino_Boards definition version 1.0.2 which contains few changes which interfere withthe existing PF-build.sh using Arduino_Boards v1.0.1, i had to add an additional variable.
I have tested PF-build.sh with Arduino_Boards v1.0.1 and v1.0.2

It creates identical firmware hex files, if I manage it to run in the same minute.
If one of the runs is off by a minute compared to the other there is just one line in the hex code different,
which is showing the compiling date, time and firmware version.