Preheat to saved temperature in load/unload during pause
When the printer is paused, use the saved_extruder_temperature directly
in preheat_or_continue, which is used by load/unload filament in the
menu. The material temperature is known is this case.

Also ensure bFilamentPreheatState is set. If the preheat is cancelled,
the temperature returns to the paused state instead of being cleared
(which would cool down the bed, detaching the print!)

Fixes #2385