Fix unexpected timeout about SD card access
Prusa-Firmware has a 32-bit millisecond counter for various timeout checks.
However, when accessing the SD card via SPI, the millisecond timer value is treated as a 16-bit timer (the maximum is 65.5 seconds).
So, if the lower 16 bits of the system's millisecond timer were close to 0xffff, the timeout was too early (in a bad case, 1 millisecond instead of 300 milliseconds).

Therefore, the counter variable was changed from 16 bits to 32 bits.