Implement xflash_rd_data_atomic to read from XFLASH
When reading from the main loop the stepper ISR is potentially running
and expects the SPI bus to be free.

As such any time we need to read from the XFLASH we need to do so
with interrupts disabled to ensure we don't read garbled data.

Introduce a new function: xflash_rd_data_atomic to perform a (small)
atomic read with interrupts disabled which is safe to use.