Perform the extra purge on M600 if a cooldown happened
Not doing the extra purge is sensible if the swap is performed fast
enough, but if an extruder cooldown happened while waiting for the user
in M600 this is no longer true and the tip might have softened.

Do an extra purge in such cases.

Add some comments and rename UnloadType::User into ::Purge to make the
distinction between the various modes clear.